pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "think-more"
version = "0.1.0"
description = "Structured reasoning constraints for LLM agents"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "jedi>=0.19",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]
mcp = [
    "mcp>=1.0",
    "pycg",
]

[project.scripts]
think-more-mcp = "think_more.mcp.server:run"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
